requirements.txt
django-bootstrap5
django-crispy-forms
django-bootstrap-icons
crispy-bootstrap5
djangorestframework
Pillow